fix the setters and make the budget a float

+++ Shop.h
@@ -7,7 +7,7 @@ class Shop
     private:
 
     std::string name;
-    int budget;
+    float budget;
     //productDB
     //clientDB
     //receiptDB
